Move folder to root
