--- skeleton/behat/behat.local.yml.example
@@ -1,11 +1,0 @@
-default:
-  extensions:
-    Behat\MinkExtension\Extension:
-      base_url: http://localhost/skeleton/www
-      goutte:
-        guzzle_parameters:
-          redirect.disable: true
-    Drupal\DrupalExtension\Extension:
-      api_driver: "drupal"
-      drupal:
-        drupal_root: "/var/www/skeleton/www"
