--- skeleton/behat/behat.local.yml.travis
@@ -1,11 +1,0 @@
-default:
-  extensions:
-    Behat\MinkExtension\Extension:
-      base_url:  http://127.0.0.1:8080
-      goutte:
-        guzzle_parameters:
-          redirect.disable: true
-    Drupal\DrupalExtension\Extension:
-      api_driver: "drupal"
-      drupal:
-        drupal_root: "/home/travis/build/zero2one/drupal7-skeleton/www"
